fix(product-service): fix availability logic, aggregation queries, and type mismatches
Switch availability checks from ProductStatus to stockQuantity, fix inventory value and
aggregation pipelines, align BigDecimal/Double types for MongoDB, and remove unused constants.

+++ product-service/src/main/java/com/stschool/ecommerce/productservice/dto/response/CategoryAveragePriceDto.java
@@ -5,8 +5,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -14,5 +12,5 @@ import java.math.BigDecimal;
 public class CategoryAveragePriceDto {
 
     private String category;
-    private BigDecimal avgPrice;
+    private Double avgPrice;
 }
